+++ src/main/java/com/mageddo/dnsproxyserver/docker/ContainerHostnameMatcher.java
@@ -11,7 +11,7 @@ import java.util.function.Predicate;
 
 import static com.mageddo.dnsproxyserver.docker.Docker.buildHostnamesFromServiceOrContainerNames;
 import static com.mageddo.dnsproxyserver.docker.Docker.findContainerHostname;
-import static com.mageddo.dnsproxyserver.docker.Docker.findHostnamesFromEnv;
+import static com.mageddo.dnsproxyserver.docker.DpsContainer.findHostnamesFromEnv;
 
 public class ContainerHostnameMatcher {
 
@@ -27,10 +27,9 @@ public class ContainerHostnameMatcher {
     return container -> {
 
       final List<Predicate<InspectContainerResponse>> predicates = List.of(
-        (it) -> host.isEqualTo(findContainerHostname(it.getConfig())),
-        (it) -> findHostnamesFromEnv(it.getConfig().getEnv()).contains(host),
-        (it) -> isRegisterContainerNames(config)
-          && buildHostnamesFromServiceOrContainerNames(it, config.getDomain()).contains(host)
+        (it) -> hostnameMatches(it, host),
+        (it) -> hostnamesEnvMatches(it, host),
+        (it) -> serviceOrContainerNameMatches(it, host, config)
       );
       for (final var predicate : predicates) {
         if (predicate.test(container)) {
@@ -43,5 +42,18 @@ public class ContainerHostnameMatcher {
 
   static boolean isRegisterContainerNames(final Config config) {
     return BooleanUtils.isTrue(config.getRegisterContainerNames());
+  }
+
+  public static boolean hostnameMatches(InspectContainerResponse c, Hostname host) {
+    return host.isEqualTo(findContainerHostname(c.getConfig()));
+  }
+
+  public static boolean hostnamesEnvMatches(InspectContainerResponse c, Hostname host) {
+    return findHostnamesFromEnv(c.getConfig().getEnv()).contains(host);
+  }
+
+  public static boolean serviceOrContainerNameMatches(InspectContainerResponse c, Hostname host, Config config) {
+    return isRegisterContainerNames(config)
+      && buildHostnamesFromServiceOrContainerNames(c, config.getDomain()).contains(host);
   }
 }